requirements.txt
Django==2.2.6
mysqlclient==1.4.4
djangorestframework==3.8.1